Add JSON serialization options to controllers, include System.Text.Json.Serialization namespace, and mark ICollection properties in entities with [JsonIgnore] to prevent circular references.

+++ Domain/Entities/ProductVariantEntity.cs
@@ -8,7 +8,7 @@ namespace VideoGameStore.Domain.Entities
 
         public long IdProduct { get; set; }
 
-        public short IdAccountType { get; set; }
+        public long IdAccountType { get; set; }
 
         public AccountTypeEntity? AccountType { get; private set; }
